sealed class test 3
